Replaced/rewrote function variant_call_lookup() with a more general function trie_lookup().  The number of globals this function depends on is decreased, and it conditionally places collected variables into an array supplied by the caller.  This allowed for the merging of multiple ret/N construction functions found in tr_utils.c and tries.c.
The result of the merge was the function build_ret_term() which
replaced construct_ret_for_call() that was contained in this file.

construct_answer_template() is added for computing the answer template
of a properly subsumed goal.

Return type of get_subgoal_ptr() is now the expected SGFrame.

get_call() implements the builtin with the same name.


git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@1677 756869e0-9e47-4bc5-b0a8-847305dc5e1c