Altered the manner in which the truth of negated subsumptive predicates are evaluated.  Originally I relied upon the interpretive timestamp-based relevant-answer collection function, but this requires the Time Stamp Indices to be present in the TST.  If the producer has completed before the collection, then these structures would have already been be reclaimed.
The new approach checks for a relevant answer using the
trie_get_return/2 builtin.  If one is found, then the literal can be
failed.  Otherwise, the producer must be checked for completion and
whether the negative-delay flag has been set.  This is performed
through a new builtin, lrd_success/2.  Therefore, slg_not/1 is restored
for handling variant tabled predicates only.


git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@2123 756869e0-9e47-4bc5-b0a8-847305dc5e1c