New Features:
DL Rules There are a couple of new predicates: retractallDLRule/3 and
retractallIntRule/3.  Also, I took any component checks out of DL
rules.  The code works (at least for an initial implementation) and
there are tests for DL rules in the test suite.  However, the API for
DL rules wont be set in stone until I finish integrating DL rules with
other aspects of the system.

Setting consistency checks.  I've added a couple of predicates,
addCheckToContext(Check,Context) and
removeCheckFromContext(Check,Context) along with other context
handling routines.  Using these predicates makes adjusting consistency
checking easier and more powerful.

You now set flags in cdf_init.P -- as a default identity checks
are made for newExtTermSingle and newExtTermBatch, but not for
anything else.  I dont have explicit tests for the checks in the test
suite but will put that in soon.

Also, I've added a scripts directory to CDF and use the script
cdfconftest.sh to test out the different combinations for tabled_isa
and XJ.  cdfconftest.sh should probably be run before checking in
changes, but you can just run the regular test out of the XSB test
directory if you just want to make sure that things are going ok.
Hopefully this will probably eliminate some of the errors that arise
when CDF is updated.



Backwards incompatabilities.

-- I had trouble with the XSB module system that made me turn
cdf_checks.P from a module to a file in usermod.  I don't think there
are more than about 3 files that directly import from cdf_checks.P,
but the import statements will have to be changed.

CDF initialization now requires a new file, cdf_init_checks.P to
initialize the checks.  This is somewhat inelegant, and I'd rather do
this in cdf_config.P, but it was difficult to do this.


git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@3834 756869e0-9e47-4bc5-b0a8-847305dc5e1c