Added check for null pointers to mem_xxxoc() functions.  Thanks to our now strict use of mem_xxxoc() functions, this means we always check for null pointers returned by malloc.  Next thing you know, I'll be rotating my tires every year.  If we do get a null pointer, we call a non-ISO xsb_memory_error().
Updated documentation in exceptions.tex, and in the course of so doing
added some new ISO-error types to error_handler.P.  There are still
corresponding functions to be added to error_xsb.c Other documentation
for setof and other accumulated changes.

The other change was to get rid of a few stray references to
resume_compl_suspension_inst2.


git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@4597 756869e0-9e47-4bc5-b0a8-847305dc5e1c