Changes to mutex handling for IO that should affect only the MT engine. Rather than having a global IO mutex, I've added a mutex for each stream, and used MUTEX_IO to lock the open_files array itself.  This change was prompted by the fact that I can't have one thread waiting on an input-stream via, say, a read and have that thread create other threads that write to the output stream and exit when done.
The result is that we no longer have contention when different
threads write to different streams.  At the same time, we end up
doing no more locking than before -- a little less, in fact due
to some manipulations in xsb_writ.P Unfortunately, the stream
mutexes are recursive (as with MUTEX_IO).  We can't get rid of
this until we rewrite write_term.

Now, back to handling memory_errors.


git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@4602 756869e0-9e47-4bc5-b0a8-847305dc5e1c