Prolog code for unify_with_occur_check
In addition, fixed bug in batched evaluation that I believe is of long standing.

In the batched fixed point, a subgoal S may be checked for fixed
point even if it has been completed and its answer return lists
have been reclaimed.  This doesn't usually happen but can if S is
the leader of the ASCC.  In such a case, the fixed point check
traverses each consumer choice point for S and determines whether
all answers have been returned to the choice point.  However, if
the answer return list for S has been reclaimed and its space
reused, the fixpoint check may mistakenly think that an answer
has not been returned to the choice point -- a dangerous situation.

The fix is simple -- a check for reclamation of the subgoal.  The
check will also be made for local evaluation, even though it
won't be needed there.


git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@4801 756869e0-9e47-4bc5-b0a8-847305dc5e1c