Fixes to get clean run of testsuite under Valgrind for uential engine.  Most of these were inconsequential and were easily fixed.  One was non-trivial and exposed an obscure memory bug.
On line 304 of heap_xsb.c is the statement

  ls_top = top_of_localstk - 256;  /* extra space for environment above top */ \

I believe that David put this in a few months ago to fix a
problem with not copying over the topmost local stack environment
when reallocating the global and local stacks.  It fixed some
memory problems, and made XSB a bit more stable.

Since the same definition of ls_top is used for garbage
collecting, 256 uninitialized variables could be encountered
during GC, which could lead to problems.  In addition it is not
impossible that other C functions create "holes" in the heap (I
suspect build_delay_list() in residual.c but could be wrong on
that).  If these holes contain uninitialized cells, we could get
core dumps on GC.

My solution is to initialize cells in the glstack at init time,
and to initialize cells *between* the top of local stack and top
of heap at realloc time.

This hypothesis may seem obscure, but I've been chasing down bugs
the last few weekends that dont seem much more obscure than
this :-)

In addition, we're now set up so that we can rerun Valgrind to
find memory bugs after large changes to XSB.


git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@4975 756869e0-9e47-4bc5-b0a8-847305dc5e1c