Added a check if redefining a predicate will release code that is currently being executed.  This catches rare but very obscure problems that are very difficult to track down.
I'm thinking of adding a warning if you redefine a predicate from a
different file.  It would require keeping for each usermod predicate
the file from which it was defined.  The flat predicate namespace of
usermod is getting to be a problem for our large and spreadout programs.
(If they'd only use modules...)


git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@6143 756869e0-9e47-4bc5-b0a8-847305dc5e1c