*Many* changes to support the windows 64-bit port.  Almost all changes are either 1. Casts to shut the compiler up when a larger integer is assigned to a shorter one, or 2. Redeclarations of int (or long) to Integer (or to size_t) so that lengths are appropriate for 64-bit ports.
A couple of other minor bugs were fixed, such as:
a) error when boxed integers are required for a result of a power operation.
b) error in simplification in subsumptive case (test of symbol type was missing
the reference to the symbol in the trie record.)
c) max-min integer specifications.
d) error in term_compare on 64-bit when integers greater the 2^32 apart were compared.
... and maybe another one or two.


git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@7030 756869e0-9e47-4bc5-b0a8-847305dc5e1c