Added ability to load_dyn a file into a module.  There are two ways:
1. Explicitly annotate rule heads with the Module: annotation. E.g., a
file full of facts such as mymod:p(1,2).  would assert that fact into
the indicated module.  For load_dync, this would have to be written as
:(mod,p(1,2)) to be in canonical form.  For a large file, this is
quite inefficient since each fact requires looking up the module, and
that takes time.

2. I've hadded a third argument to load_dyn/2 and load_dync/2 to take
a module name.  E.g. load_dync(big_data,1,mymod). (The second arg is
direction, here to use assertz).  Here all structures are put into
mymod.  This works fine for datalog facts, loading them into the
indicated module.  But structured terms will also be put in that
module, which may not be what is desired.  We have to see if this is
useful.  (Changing it to put terms in usermod and predicates into the
indicated module would take some work.)


git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@7109 756869e0-9e47-4bc5-b0a8-847305dc5e1c