XSB keeps track of choice points in two ways.  1) by running the choice point stack (for table and clause gc).  2) by maintaining a "visitors" field in tabled subgoal frames to determine when a given completed table is being backtracked through (used by incremental tabling).  The visitors field is incremented when a completed table is called, and decremented when backtracking out of the table.  In addition, if the choice point is cut over or thrown over by an exception, the visitors field is similarly decremented.
However when a table T is early completed the breg is reset to
the choice point for T, cutting over intermediate choice points.
This seems correct (as opposed to simply resetting the choice
point for T but not breg), but in this case the cut-over choice
points need to be examined so that their visitors field is
decremented.  In addition, the choice points need to be	checked
to perform any call cleanup.

To fix this I put a simple while loop to check for choice points
that need to be handled	   before resetting the breg after
performing early completion.


git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@7506 756869e0-9e47-4bc5-b0a8-847305dc5e1c