Small cleanup of abolish_table_call_internal.  This predicate backtracked thrugh all calls that unify with a given input term, abolishing each one. To do this, it uses get_calls, which collects a list of pointers to such tables.
However if one of the tables on the list depends on another via the SLG delay list, and if transitive abolishes are done, the latter table
may be abolished, while a pointer to it is still in the list.  So, with transitive abolishes, each pointer must be rechecked to make sure its safe.





git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@8255 756869e0-9e47-4bc5-b0a8-847305dc5e1c