Major change to	transitive abolishes for calls.  As before, if a table T_a is abolished, any table	T' that	has an answer that depends on T_a or an answer in T_a is abolished.  Now, any table T'' upon which T_a depends (i.e., T_a has an answer that has T'' or an answer in T'' in its delay list) is affected in the following manner.  Any backpointers from T'' to T_a are removed from the subgoal or answer PNDEs.  This	means that after table abolishes, all backpointers from non-abolished tables will remain valid.  (At least, I sure hope so.)
In the first version of	transitive abolishes, tables such as T''
were unaffected.  However, this left invalid back pointers and
led to core dumps when running Ergo.  The next version that was
has been in place for a couple of weeks actually abolished tables
such as T'', but this was too extreme.	So hopefully this version
will be just right.

The above changes were in tr_utils.c.  The slgdelay* changes were so that remove_pndes could be  
used in tr_utils.c



git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@8275 756869e0-9e47-4bc5-b0a8-847305dc5e1c