Changes to allow forest logging to be turned on or off on a per-predicate basis.
In addition, I restructured the PSC structure to include bitmapped fields.  It turned out that this wasnt necessary for the forest 
logging change, as I keep the forest logging bit in the TIF, but was a good change for other future modifications that we may want to 
make to PSC structures, as well as making it a little easier to support the combination of incremental and interned tables.



git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@8303 756869e0-9e47-4bc5-b0a8-847305dc5e1c