Previous adding of cast in tab_structs.h broke.  Interestingly the smaller (level_num) needed to be cast to the larger (flags{..]) since the flag really held a number greater than 32 bits.)
Added CYGWIN test to allow definition of PTHREAD_MUTEX_ERRORCHECK_NP
and friends so cygwin compilation could get further.  Teri has a
comment just before this change (line 217 in thread_xsb.h) which
sounds relevant, but I don't understand it.  Is this what was meant?
Or was it to say that this change of mine is probably wrong?



git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@8542 756869e0-9e47-4bc5-b0a8-847305dc5e1c