sym = insert_psc(ptoc_string(CTXTc 1),(int)ptoc_int(CTXTc 2), global_mod, &value);           init_psc_ep_info(pair_psc(sym)); // reset to reload 	         psc_set_data(pair_psc(sym),(Psc)makestring(string_find(mod_name,1))); 		 +      //      psc_set_env(tpsc,T_UNLOADED); 		 +      if (get_type(pair_psc(sym)) == 0) 		 +       psc_set_type(pair_psc(sym),T_UDEF); 		      } else { 		             mod_psc = pair_psc(insert_module(0, mod_name)); 			            sym = insert_psc(ptoc_string(CTXTc 1), (int)ptoc_int(CTXTc 2), mod_psc, &value);
Added the 3 +'d lines above to psc_import to make the import directive
consistent with the import declaration as executed by the loader.  The
motivation for this is to make predicates that are imported but
undefined have the type T_UDEF, which can be reflected as unloaded by predicate_property.



git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@9189 756869e0-9e47-4bc5-b0a8-847305dc5e1c