Changed interpreter so that querying:
?- goal.

with spaces between the . and the newline will act the same as if there
were no spaces.  Before, if there were spaces, the goal succeeded, but the interpreter said 'no'.

Perhaps because I use emacs, this happened to me occasionally, and I finally
figured out what the problem was and fixed it.